refact(services-mutation): dividing the mutation into more components
