+++ dashboard/src/features/services/dialogs/index.ts
@@ -1,2 +1,2 @@
-export * from './delete-confirmation.dialog'
-export * from './mutation.dialog'
+export * from "./delete-confirmation.dialog";
+export * from "./mutation";
